moved create function to __init__
